backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
web3
python-dotenv
pydantic==2.7.0
sqlalchemy==2.0.30
